requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.2.7
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.8.0
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.3
django-watson==1.6.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
inflection==0.5.1
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
moviepy==1.0.3
numpy==1.21.2
packaging==21.0
parso==0.8.2
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.3.2
proglog==0.1.9
prompt-toolkit==3.0.20
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
traitlets==5.1.0
transliterate==1.10.2
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
xlrd==2.0.1
